requirements.txt
google-adk
rich
PyMuPDF
